+++ src/test/java/kitchenpos/menu/domain/MenuNameTest.java
@@ -1,29 +1,29 @@
-package kitchenpos.domain.product;
+package kitchenpos.menu.domain;
 
 import kitchenpos.BaseTest;
-import kitchenpos.product.domain.ProductName;
+import kitchenpos.menu.domain.MenuName;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class ProductNameTest extends BaseTest {
+class MenuNameTest extends BaseTest {
 
     @Test
-    void 상품_이름을_생성한다() {
+    void 메뉴_이름을_생성한다() {
         // given
         String name = "pizza";
 
         // when, then
         Assertions.assertThatNoException()
-                .isThrownBy(() -> new ProductName(name));
+                .isThrownBy(() -> new MenuName(name));
     }
 
     @Test
-    void 상품_이름은_공백일_수_있다() {
+    void 메뉴_이름은_공백일_수_있다() {
         // given
         String name = "";
 
         // when, then
         Assertions.assertThatNoException()
-                .isThrownBy(() -> new ProductName(name));
+                .isThrownBy(() -> new MenuName(name));
     }
 }
