+++ src/test/java/kitchenpos/product/domain/ProductNameTest.java
@@ -1,29 +1,29 @@
-package kitchenpos.domain.menu;
+package kitchenpos.product.domain;
 
 import kitchenpos.BaseTest;
-import kitchenpos.menu.domain.MenuName;
+import kitchenpos.product.domain.ProductName;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class MenuNameTest extends BaseTest {
+class ProductNameTest extends BaseTest {
 
     @Test
-    void 메뉴_이름을_생성한다() {
+    void 상품_이름을_생성한다() {
         // given
         String name = "pizza";
 
         // when, then
         Assertions.assertThatNoException()
-                .isThrownBy(() -> new MenuName(name));
+                .isThrownBy(() -> new ProductName(name));
     }
 
     @Test
-    void 메뉴_이름은_공백일_수_있다() {
+    void 상품_이름은_공백일_수_있다() {
         // given
         String name = "";
 
         // when, then
         Assertions.assertThatNoException()
-                .isThrownBy(() -> new MenuName(name));
+                .isThrownBy(() -> new ProductName(name));
     }
 }
